feat(coreInfo): centralise storage of image assets
The root json object now has an 'a' key with a list of bytes objects. Each image then stores an index of this list.

This allows duplicate images/assets to only be stored once (e.g. when a pdf is stored multiple times for each page in the pdf).

Note that duplicity doesn't increase memory usage since we're just using the same object (actually stored only once in memory) but in multiple variables.