Created Step 0 in Line 182 strings.i18n.yaml (#927)
* Created Step 0 in Line 182 strings.i18n.yaml

Fixed the instructions regarding "How do I change my encryption password?" in Line 180.

The old instructions make a crucial mistake of saying "do this", and afterwards saying "do this first".

I propose to create Step 0. This indicates, that this is an optional prerequisite, but should at least be read before simply logging out. Thus, the possibility of reading this instruction and missing this "backup-step" is lowered.

* chore: shorten step 0

* i18n: apply strings

---------

Co-authored-by: Adil Hanney <adilh+github@orchid.anonaddy.com>